pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["diivanovsky <diivanovsky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
prospector = "^1.10.2"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
djoser = "^2.2.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
